Move QuestionDefinitions to new file
Inline methods

Change DatabaseSeedController.seed to DatabaseSeedController.seedPrograms

Canonical questions -> sample questions

Mock/Simple -> Large/Small

Clean up DATE_QUESTION_DEFINITION

Change question names

Update Name question name

Fmt fix

Move around files with new seeding package
