+++ server/test/tasks/DatabaseSeedTaskTest.java
@@ -2,6 +2,7 @@ package tasks;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import controllers.dev.seeding.DatabaseSeedTask;
 import java.util.Set;
 import models.Question;
 import org.junit.Before;
@@ -27,7 +28,7 @@ public class DatabaseSeedTaskTest extends ResetPostgres {
   }
 
   @Test
-  public void seedCanonicalQuestions_whenQuestionsNotSeededYet_itSeedsTheCanonicalQuestions()
+  public void seedQuestions_whenQuestionsNotSeededYet_itSeedsTheCanonicalQuestions()
       throws Exception {
     assertThat(getAllQuestions().size()).isEqualTo(0);
 
@@ -42,7 +43,7 @@ public class DatabaseSeedTaskTest extends ResetPostgres {
   }
 
   @Test
-  public void seedCanonicalQuestions_whenSomeQuestionsAlreadySeeded_itSeedsTheMissingOnes() {
+  public void seedQuestions_whenSomeQuestionsAlreadySeeded_itSeedsTheMissingOnes() {
     instanceOf(QuestionService.class)
         .create(
             new DateQuestionDefinition(
